--- packages/ui/src/utils/constants.js
@@ -1,625 +1,0 @@
-import searchIcon from "../assets/searchIcon.svg";
-import curvedArrow from "../assets/curvedArrow.svg";
-import rapidLogo from "../assets/rapid.svg";
-import searchLogo from "../assets/search.svg";
-import databaseLogo from "../assets/database.svg";
-import dragAndDropBg from "../assets/DragAndDropBg.svg";
-import notion from "../assets/notion.png";
-import slack from "../assets/slack.png";
-import google_drive from "../assets/google_drive.png";
-import whatsapp from "../assets/whatsapp.png";
-import g_calendar from "../assets/g_calendar.png";
-import mailchimp from "../assets/mailchimp.png";
-import zapier from "../assets/zapier.png";
-import zendesk from "../assets/zendesk.png";
-import stripe from "../assets/stripe.png";
-import jira from "../assets/jira.png";
-import intercom from "../assets/intercom.png";
-import figma from "../assets/figma.png";
-import discord from "../assets/discord.png";
-import bitbucket from "../assets/bitbucket.png";
-import confluence from "../assets/confluence.png";
-import dropbox from "../assets/dropbox.png";
-
-export const SVGS = {
-  searchIcon,
-  curvedArrow,
-  amazon: "https://upload.wikimedia.org/wikipedia/commons/4/4a/Amazon_icon.svg",
-  apple:
-    "https://upload.wikimedia.org/wikipedia/commons/f/fa/Apple_logo_black.svg",
-  adobe:
-    "https://upload.wikimedia.org/wikipedia/commons/8/8d/Adobe_Corporate_Logo.png",
-  google:
-    "https://www.google.com/images/branding/googlelogo/2x/googlelogo_color_272x92dp.png",
-  microsoft:
-    "https://upload.wikimedia.org/wikipedia/commons/9/96/Microsoft_logo_%282012%29.svg",
-  dragAndDropBg,
-};
-
-export const COMPANIES = [
-  {
-    id: 1,
-    name: "Amazon",
-    logo: SVGS.amazon,
-  },
-  {
-    id: 2,
-    name: "Apple",
-    logo: SVGS.apple,
-  },
-  {
-    id: 3,
-    name: "Adobe",
-    logo: SVGS.adobe,
-  },
-  {
-    id: 4,
-    name: "Google",
-    logo: SVGS.google,
-  },
-  {
-    id: 5,
-    name: "Microsoft",
-    logo: SVGS.microsoft,
-  },
-];
-
-export const featureItems = [
-  {
-    icon: databaseLogo,
-    title: "Comprehensive Database",
-    content:
-      "Tap into a vast logo library with thousands of brands, continuously refreshed and expanding.",
-  },
-  {
-    icon: searchLogo,
-    title: "Customizable Search Insights",
-    content:
-      "Gain insights on search patterns to spot missing logos and keep collections comprehensive.",
-  },
-  {
-    icon: rapidLogo,
-    title: "Fast & Reliable API Access",
-    content:
-      "Access logos instantly with fast, dependable APIs built to minimize downtime and maximize efficiency.",
-  },
-];
-
-export const HEADER_ITEMS = [
-  {
-    name: "home",
-    title: "Home",
-    url: "/",
-  },
-  {
-    name: "docs",
-    title: "Docs",
-    url: "/docs",
-  },
-  {
-    name: "features",
-    title: "Features",
-    url: "/features",
-  },
-  {
-    name: "pricing",
-    title: "Pricing",
-    url: "/pricing",
-  },
-  {
-    name: "about",
-    title: "About",
-    url: "/about",
-  },
-];
-export const FOOTER_ITEMS = [
-  {
-    name: "home",
-    title: "Home",
-    url: "/",
-  },
-  {
-    name: "docs",
-    title: "Docs",
-    url: "/",
-  },
-  {
-    name: "demo",
-    title: "Demo",
-    url: "/demo",
-  },
-  {
-    name: "features",
-    title: "Features",
-    url: "/features",
-  },
-  {
-    name: "pricing",
-    title: "Pricing",
-    url: "/pricing",
-  },
-  {
-    name: "about",
-    title: "About",
-    url: "/about",
-  },
-  {
-    name: "privacy",
-    title: "Privacy",
-    url: "/privacy",
-  },
-];
-export const companies = [
-  {
-    id: "A1B2C3D4E5",
-    companyImage: "Amazon.png",
-    createdAt: "15 Sep 2023",
-    updatedAt: "17 Sep 2023",
-  },
-  {
-    id: "F6G7H8I9J0",
-    companyImage: "Apple.png",
-    createdAt: "01 Oct 2024",
-    updatedAt: "12 Oct 2024",
-  },
-  {
-    id: "K1L2M3N4O5",
-    companyImage: "Google.png",
-    createdAt: "11 Aug 2024",
-    updatedAt: "12 Aug 2024",
-  },
-  {
-    id: "P6Q7R8S9T0",
-    companyImage: "Microsoft.png",
-    createdAt: "22 Jul 2023",
-    updatedAt: "23 Jul 2023",
-  },
-  {
-    id: "U1V2W3X4Y5",
-    companyImage: "Tesla.png",
-    createdAt: "09 Feb 2024",
-    updatedAt: "18 Feb 2024",
-  },
-  {
-    id: "Z6A7B8C9D0",
-    companyImage: "Meta.png",
-    createdAt: "13 Dec 2023",
-    updatedAt: "15 Dec 2023",
-  },
-  {
-    id: "E1F2G3H4I5",
-    companyImage: "JohnsonAndJohnson.png",
-    createdAt: "27 Nov 2024",
-    updatedAt: "30 Nov 2024",
-  },
-  {
-    id: "J6K7L8M9N0",
-    companyImage: "JPMorganChase.png",
-    createdAt: "04 Apr 2024",
-    updatedAt: "08 Apr 2024",
-  },
-  {
-    id: "O1P2Q3R4S5",
-    companyImage: "Walmart.png",
-    createdAt: "19 Jun 2024",
-    updatedAt: "23 Jun 2024",
-  },
-  {
-    id: "T6U7V8W9X0",
-    companyImage: "BerkshireHathaway.png",
-    createdAt: "01 Mar 2024",
-    updatedAt: "10 Mar 2024",
-  },
-  {
-    id: "Y1Z2A3B4C5",
-    companyImage: "Disney.png",
-    createdAt: "10 Jan 2024",
-    updatedAt: "12 Jan 2024",
-  },
-  {
-    id: "D6E7F8G9H0",
-    companyImage: "Intel.png",
-    createdAt: "30 Aug 2023",
-    updatedAt: "03 Sep 2023",
-  },
-  {
-    id: "I1J2K3L4M5",
-    companyImage: "Nike.png",
-    createdAt: "14 Jun 2024",
-    updatedAt: "20 Jun 2024",
-  },
-  {
-    id: "N6O7P8Q9R0",
-    companyImage: "PepsiCo.png",
-    createdAt: "23 Feb 2024",
-    updatedAt: "25 Feb 2024",
-  },
-  {
-    id: "S1T2U3V4W5",
-    companyImage: "CocaCola.png",
-    createdAt: "29 Oct 2023",
-    updatedAt: "01 Nov 2023",
-  },
-  {
-    id: "X6Y7Z8A9B0",
-    companyImage: "ProcterAndGamble.png",
-    createdAt: "05 May 2024",
-    updatedAt: "06 May 2024",
-  },
-  {
-    id: "C1D2E3F4G5",
-    companyImage: "Pfizer.png",
-    createdAt: "09 Dec 2023",
-    updatedAt: "11 Dec 2023",
-  },
-  {
-    id: "H6I7J8K9L0",
-    companyImage: "GeneralElectric.png",
-    createdAt: "18 Mar 2024",
-    updatedAt: "20 Mar 2024",
-  },
-  {
-    id: "M1N2O3P4Q5",
-    companyImage: "IBM.png",
-    createdAt: "07 Apr 2024",
-    updatedAt: "09 Apr 2024",
-  },
-  {
-    id: "R6S7T8U9V0",
-    companyImage: "Verizon.png",
-    createdAt: "16 Sep 2023",
-    updatedAt: "20 Sep 2023",
-  },
-  {
-    id: "W1X2Y3Z4A5",
-    companyImage: "AT&T.png",
-    createdAt: "10 Aug 2024",
-    updatedAt: "13 Aug 2024",
-  },
-  {
-    id: "B6C7D8E9F0",
-    companyImage: "Oracle.png",
-    createdAt: "24 Oct 2024",
-    updatedAt: "28 Oct 2024",
-  },
-  {
-    id: "G1H2I3J4K5",
-    companyImage: "Cisco.png",
-    createdAt: "15 Jul 2024",
-    updatedAt: "18 Jul 2024",
-  },
-  {
-    id: "L6M7N8O9P0",
-    companyImage: "Netflix.png",
-    createdAt: "25 Feb 2024",
-    updatedAt: "01 Mar 2024",
-  },
-  {
-    id: "Q1R2S3T4U5",
-    companyImage: "ExxonMobil.png",
-    createdAt: "19 Nov 2023",
-    updatedAt: "21 Nov 2023",
-  },
-  {
-    id: "V6W7X8Y9Z0",
-    companyImage: "Chevron.png",
-    createdAt: "22 Dec 2023",
-    updatedAt: "27 Dec 2023",
-  },
-  {
-    id: "A1B2C3D4E6",
-    companyImage: "GoldmanSachs.png",
-    createdAt: "14 Jan 2024",
-    updatedAt: "18 Jan 2024",
-  },
-  {
-    id: "F6G7H8I9J1",
-    companyImage: "WellsFargo.png",
-    createdAt: "05 Mar 2024",
-    updatedAt: "09 Mar 2024",
-  },
-  {
-    id: "K1L2M3N4O6",
-    companyImage: "BankOfAmerica.png",
-    createdAt: "06 Oct 2024",
-    updatedAt: "12 Oct 2024",
-  },
-  {
-    id: "P6Q7R8S9T1",
-    companyImage: "Target.png",
-    createdAt: "18 Jun 2024",
-    updatedAt: "21 Jun 2024",
-  },
-  {
-    id: "U1V2W3X4Y6",
-    companyImage: "Costco.png",
-    createdAt: "12 Apr 2024",
-    updatedAt: "15 Apr 2024",
-  },
-  {
-    id: "Z6A7B8C9D1",
-    companyImage: "HomeDepot.png",
-    createdAt: "03 Jul 2024",
-    updatedAt: "09 Jul 2024",
-  },
-  {
-    id: "E1F2G3H4I6",
-    companyImage: "UPS.png",
-    createdAt: "21 Aug 2023",
-    updatedAt: "25 Aug 2023",
-  },
-  {
-    id: "J6K7L8M9N1",
-    companyImage: "FedEx.png",
-    createdAt: "28 Nov 2024",
-    updatedAt: "02 Dec 2024",
-  },
-  {
-    id: "O1P2Q3R4S6",
-    companyImage: "McDonalds.png",
-    createdAt: "01 Jun 2024",
-    updatedAt: "03 Jun 2024",
-  },
-  {
-    id: "T6U7V8W9X1",
-    companyImage: "Starbucks.png",
-    createdAt: "07 Feb 2024",
-    updatedAt: "12 Feb 2024",
-  },
-  {
-    id: "Y1Z2A3B4C6",
-    companyImage: "Visa.png",
-    createdAt: "20 May 2024",
-    updatedAt: "24 May 2024",
-  },
-  {
-    id: "D6E7F8G9H1",
-    companyImage: "Mastercard.png",
-    createdAt: "17 Apr 2024",
-    updatedAt: "22 Apr 2024",
-  },
-  {
-    id: "I1J2K3L4M6",
-    companyImage: "AmericanExpress.png",
-    createdAt: "04 Mar 2024",
-    updatedAt: "10 Mar 2024",
-  },
-  {
-    id: "N6O7P8Q9R1",
-    companyImage: "GeneralMotors.png",
-    createdAt: "11 Oct 2024",
-    updatedAt: "18 Oct 2024",
-  },
-  {
-    id: "S1T2U3V4W6",
-    companyImage: "Ford.png",
-    createdAt: "30 Apr 2024",
-    updatedAt: "07 May 2024",
-  },
-  {
-    id: "X6Y7Z8A9B1",
-    companyImage: "Boeing.png",
-    createdAt: "13 Aug 2024",
-    updatedAt: "20 Aug 2024",
-  },
-  {
-    id: "C1D2E3F4G6",
-    companyImage: "LockheedMartin.png",
-    createdAt: "29 Sep 2023",
-    updatedAt: "06 Oct 2023",
-  },
-  {
-    id: "H6I7J8K9L1",
-    companyImage: "NorthropGrumman.png",
-    createdAt: "05 Dec 2023",
-    updatedAt: "10 Dec 2023",
-  },
-  {
-    id: "M1N2O3P4Q6",
-    companyImage: "Raytheon.png",
-    createdAt: "12 Nov 2024",
-    updatedAt: "15 Nov 2024",
-  },
-  {
-    id: "R6S7T8U9V1",
-    companyImage: "Samsung.png",
-    createdAt: "18 May 2024",
-    updatedAt: "25 May 2024",
-  },
-  {
-    id: "W1X2Y3Z4A6",
-    companyImage: "Sony.png",
-    createdAt: "21 Jan 2024",
-    updatedAt: "28 Jan 2024",
-  },
-  {
-    id: "B6C7D8E9F1",
-    companyImage: "Panasonic.png",
-    createdAt: "03 Oct 2023",
-    updatedAt: "11 Oct 2023",
-  },
-  {
-    id: "G1H2I3J4K6",
-    companyImage: "LG.png",
-    createdAt: "10 Feb 2024",
-    updatedAt: "18 Feb 2024",
-  },
-  {
-    id: "L6M7N8O9P1",
-    companyImage: "Huawei.png",
-    createdAt: "12 Jun 2024",
-    updatedAt: "20 Jun 2024",
-  },
-  {
-    id: "Q1R2S3T4U6",
-    companyImage: "Dell.png",
-    createdAt: "17 Aug 2023",
-    updatedAt: "23 Aug 2023",
-  },
-  {
-    id: "V6W7X8Y9Z1",
-    companyImage: "HP.png",
-    createdAt: "14 Apr 2024",
-    updatedAt: "20 Apr 2024",
-  },
-  {
-    id: "A1B2C3D4E7",
-    companyImage: "Lenovo.png",
-    createdAt: "09 Mar 2024",
-    updatedAt: "15 Mar 2024",
-  },
-  {
-    id: "F6G7H8I9J2",
-    companyImage: "Xiaomi.png",
-    createdAt: "18 Nov 2023",
-    updatedAt: "25 Nov 2023",
-  },
-  {
-    id: "K1L2M3N4O7",
-    companyImage: "Philips.png",
-    createdAt: "21 Dec 2023",
-    updatedAt: "28 Dec 2023",
-  },
-  {
-    id: "P6Q7R8S9T2",
-    companyImage: "Siemens.png",
-    createdAt: "05 Jan 2024",
-    updatedAt: "12 Jan 2024",
-  },
-  {
-    id: "U1V2W3X4Y7",
-    companyImage: "Honeywell.png",
-    createdAt: "27 Apr 2024",
-    updatedAt: "05 May 2024",
-  },
-];
-
-export const INFO = [
-  { title: "Users", api: 5.2 },
-  { title: "Keys", api: -6.8 },
-  { title: "Hits", api: 9.2 },
-  { title: "Requests", api: -5.3 },
-];
-
-export const PASSWORD_VALIDATION_MESSAGES = {
-  required: "Password is required!",
-  minLength: (minLength) =>
-    `Password must be at least ${minLength} characters long.`,
-  maxLength: (maxLength) =>
-    `Password must be no more than ${maxLength} characters long.`,
-  uppercase: "Password must contain at least one uppercase letter.",
-  lowercase: "Password must contain at least one lowercase letter.",
-  digit: "Password must contain at least one digit.",
-  specialChar: "Password must contain at least one special character.",
-  generalError: "Password does not meet security requirements.",
-};
-
-export const INTEGRATIONS = [
-  { id: 1, src: notion, alt: "Notion" },
-  { id: 2, src: slack, alt: "Slack" },
-  { id: 3, src: google_drive, alt: "Google Drive" },
-  { id: 4, src: intercom, alt: "Intercom" },
-  { id: 5, src: stripe, alt: "Stripe" },
-  { id: 6, src: dropbox, alt: "Dropbox" },
-  { id: 7, src: jira, alt: "Jira" },
-  { id: 8, src: zapier, alt: "Zapier" },
-  { id: 9, src: figma, alt: "Figma" },
-  { id: 10, src: confluence, alt: "Confluence" },
-  { id: 11, src: mailchimp, alt: "Mailchimp" },
-  { id: 12, src: zendesk, alt: "Zendesk" },
-  { id: 13, src: g_calendar, alt: "Google Calendar" },
-  { id: 14, src: whatsapp, alt: "WhatsApp" },
-  { id: 15, src: discord, alt: "Discord" },
-  { id: 16, src: bitbucket, alt: "Bitbucket" },
-];
-
-export const ABOUT_TEXT = {
-  DESCRIPTION:
-    "From startups to enterprises, our platform offers an extensive collection of company logos, enabling smooth integration and consistent branding. Our APIs are designed to make logo retrieval effortless, providing scalable solutions that adapt to your business's evolving branding requirements.",
-};
-
-export const FAQ = [
-  {
-    question: "Why should I use Openlogo?",
-    answer:
-      "Openlogo makes it easy to retrieve logos quickly and reliably with seamless API integration, flexible pricing, and scalable access to a growing logo library. Customize logo sizes and formats to fit your needs effortlessly.",
-  },
-  {
-    question: "How do I get started with Openlogo?",
-    answer:
-      "Getting started is easy! Sign up to create your account and generate a unique API key. With your key, explore our API features to retrieve logos in different formats (e.g., PNG, SVG) and sizes tailored to your needs. Use our comprehensive documentation to guide your integration and testing process.",
-  },
-  {
-    question: "What payment methods do you support?",
-    answer:
-      "Currently, our plans are in a free trial period, and you can use the API at no cost. We're continuously expanding our library by adding new logos every day. In the future, we plan to introduce a variety of payment methods to suit your convenience.",
-  },
-];
-
-const PLANS = [
-  {
-    index: 0,
-    name: "HOBBY",
-    pricing: 0,
-    tagline: "Try for free for individuals",
-    keypoints: [
-      "Fortune 500 company logo",
-      "500 API calls per month",
-      "2 API keys",
-      "Basic analytics",
-      "48-72 hour of response time",
-    ],
-  },
-  {
-    index: 1,
-    name: "PRO",
-    pricing: 1500,
-    tagline: "Grow with pro plan.",
-    keypoints: [
-      "Hobby plan plus",
-      "10000 API calls per month",
-      "5 API keys",
-      "Advance analytics",
-      "12-36 hours of response time",
-    ],
-  },
-  {
-    index: 2,
-    name: "CUSTOM PRICING",
-    pricing: null,
-    tagline: "Large teams with unlimited users.",
-    keypoints: [
-      "Fortune 500 company logo + unlimited private logos",
-      "Unlimited API calls per month",
-      "50 API keys",
-      "Advance analytics",
-      "Priority support",
-    ],
-  },
-];
-
-export default PLANS;
-
-export const SETTING = [
-  {
-    title: "Download account data",
-    subtitle: "Download your account data and move to other device with ease.",
-    buttontitle: "Download",
-  },
-  {
-    title: "Delete account",
-    subtitle:
-      "This will permanently delete your account and all associated data.",
-    buttontitle: "Delete Account",
-  },
-];
-
-export const HAMBURGER = {
-  src: "hamburger.svg",
-  alt: "Hamburger Icon",
-};
-
-export const CROSS = {
-  src: "close-icon.svg",
-  alt: "Close Icon",
-};
